# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Rental Api",
        "description": "Descrição da Sua API",
        "contact": {
            "email": "willnmafra@gmail.com"
        },
        "license": {
            "name": "MIT License"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/car-model": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Get list of car models",
                "operationId": "6712dd60eb0c3386dfb34f1d81f39a0d",
                "parameters": [
                    {
                        "name": "columns",
                        "in": "query",
                        "description": "(optional)Columns to select from car models table. ex: /api/car?columns=brand_id,image,doors,air_bag,abs",
                        "schema": {
                            "description": "List of columns separated by comma",
                            "type": "string",
                            "example": "brand_id,name,id"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Filter for brands table",
                        "schema": {
                            "properties": {
                                "brand_name": {
                                    "type": "text",
                                    "example": "ilike:%ford%"
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "paginate",
                        "in": "query",
                        "description": "Number of items per page for pagination",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Store a newcar model",
                "operationId": "75bd453712f3f7fed32025281562c3c9",
                "requestBody": {
                    "description": "Car model data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Car Model Name",
                                        "type": "string",
                                        "example": "320i"
                                    },
                                    "brand_id": {
                                        "description": "Brand Id",
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "doors": {
                                        "description": "Number of doors",
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "seats": {
                                        "description": "Number of seats",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "abs": {
                                        "description": "The car have abs? (1 to true and 0 to false)",
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "air_bag": {
                                        "description": "The car have air bag? (1 to true and 0 to false)",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "image": {
                                        "description": "Image file of the car model",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Json with car model data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Car Model xNamex created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/car-model/{carModelId}": {
            "get": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Get details of a specific car model",
                "operationId": "b1ffde9018d96a36e8c4fdd5a515c314",
                "parameters": [
                    {
                        "name": "carModelId",
                        "in": "path",
                        "description": "ID of the car model",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Car Model not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Update a car model",
                "operationId": "13cf19c2d6a6c3548688b55aa1cb4dcc",
                "parameters": [
                    {
                        "name": "carModelId",
                        "in": "path",
                        "description": "ID of the car model",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Car model data",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "description": "method for the request",
                                        "type": "string",
                                        "example": "patch"
                                    },
                                    "name": {
                                        "description": "Car Model Name",
                                        "type": "string",
                                        "example": "i30"
                                    },
                                    "brand_id": {
                                        "description": "Brand Id",
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "doors": {
                                        "description": "Number of doors",
                                        "type": "integer",
                                        "example": "4"
                                    },
                                    "seats": {
                                        "description": "Number of seats",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "abs": {
                                        "description": "The car have abs? (1 to true and 0 to false)",
                                        "type": "integer",
                                        "example": "0"
                                    },
                                    "air_bag": {
                                        "description": "The car have air bag? (1 to true and 0 to false)",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "image": {
                                        "description": "Image file of the car model",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Brand updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Car Models"
                ],
                "summary": "Delete a car model",
                "operationId": "581c1f0113ba5b95178a620eee5ef391",
                "parameters": [
                    {
                        "name": "carModelId",
                        "in": "path",
                        "description": "ID of the car model",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "Car Model deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car Model not found"
                    }
                }
            }
        },
...
    }
}